Add Result pattern, fix profile images, posts

+++ src/Services/PostService.php
@@ -11,7 +11,6 @@ use Mailer\Mailer;
 use Models\User;
 use E_SEND_MAIL_RETURN_CODES;
 use Interfaces\IPostRepository;
-use Helpers;
 
 readonly class PostService {
     public function __construct(private IPostRepository $postRepo) {
@@ -32,12 +31,21 @@ readonly class PostService {
         return $this->postRepo->findUserPostsByUserId($userId);
     }
 
-    public function updatePost(int $postId, string $header, string $text): void {
+    public function updatePost(int $userId, int $postId, string $header, string $text): void {
         if(empty($header) || strlen($header) < 3 || strlen($header) > 64) {
             throw new Exception("Header length does not correct");    
         }
         if(empty($text) || strlen($text) < 10 || strlen($text) > 300) {
             throw new Exception("Main text length does not correct");    
+        }
+
+        $post = $this->postRepo->findPostById($postId);
+        if ($post == null) {
+            throw new Exception("Post not found");
+        }
+
+        if ($post['user_id'] != $userId) {
+            throw new Exception("You don't have permission to edit this post");
         }
 
         $rowCount = $this->postRepo->updatePostData($postId, $header, $text);
@@ -47,7 +55,16 @@ readonly class PostService {
         }
     }
 
-    public function deletePost(int $postId): void {
+    public function deletePost(int $userId, int $postId): void {
+        $post = $this->postRepo->findPostById($postId);
+        if ($post == null) {
+            throw new Exception("Post not found");
+        }
+
+        if ($post['user_id'] != $userId) {
+            throw new Exception("You don't have permission to edit this post");
+        }
+
         $rowCount = $this->postRepo->deletePostById($postId);
 
         if($rowCount == 0) {
